Added docstrings, refactored, and deleted aggregates from data fetching methods
I'm pretty sure I was using the aggregates wrong, and I am therefore not using
them anymore.

I changed the names of the parse_* functions to format_*, because that's a
better description of what they actually do.

I also made the __return_paged_data method of fetcher.Fetcher a function.