Replace OECD references with HIC, and fix correlation example
`OEC` is no longer a valid value of `incomelevel` (see https://github.com/OliverSherouse/wbdata/issues/28), so this replaces `OEC` with `HIC`.

When testing this I also discovered that the current example correlating GDP/capita and the Ease of doing business index fails due to the former being available through 2017, and the latter only being available for 2018. So this also adds some code to merge the 2018 business index with the 2017 GDP/capita to get the correlation working.